requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Brotli==1.1.0
cffi==1.17.1
click==8.1.8
cssselect2==0.8.0
exceptiongroup==1.3.0
fastapi==0.116.0
fonttools==4.58.5
greenlet==3.2.3
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
pdfcrowd==6.5.1
pdfkit==1.0.0
pillow==11.3.0
playwright==1.53.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
pyee==13.0.0
pyphen==0.17.2
sniffio==1.3.1
starlette==0.46.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
